+++ CHANGELOG.md
@@ -12,6 +12,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Add default resource to init container.
 
 ### Changed
+* Changed headless svc name to a templated name that includes the mysql cluster name
+
 ### Removed
 ### Fixed
 * Avoid set read_only conflict when graceful takeover